move CardHotkeyBadge to compoennts
