+++ apps/ligretto-frontend/src/components/blocks/game/CardHotkeyBadge/CardHotkeyBadge.tsx
