+++ apps/ligretto-frontend/src/components/blocks/game/CardHotkeyBadge/index.ts
